Moved LocalMessageBus package
