+++ spring-xd-dirt/src/test/java/org/springframework/xd/dirt/config/LocalSingleNodeInitializationTests.java
@@ -26,7 +26,7 @@ import org.junit.Test;
 
 import org.springframework.xd.dirt.cluster.ContainerAttributes;
 import org.springframework.xd.dirt.cluster.AdminAttributes;
-import org.springframework.xd.dirt.integration.bus.LocalMessageBus;
+import org.springframework.xd.dirt.integration.bus.local.LocalMessageBus;
 import org.springframework.xd.dirt.integration.bus.MessageBus;
 import org.springframework.xd.dirt.server.ApplicationUtils;
 import org.springframework.xd.dirt.server.ContainerRegistrar;
