+++ spring-xd-messagebus-local/src/main/resources/META-INF/spring-xd/bus/local-bus.xml
@@ -4,7 +4,7 @@
 	xsi:schemaLocation="http://www.springframework.org/schema/integration http://www.springframework.org/schema/integration/spring-integration.xsd
 		http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
 
-	<bean id="messageBus" class="org.springframework.xd.dirt.integration.bus.LocalMessageBus">
+	<bean id="messageBus" class="org.springframework.xd.dirt.integration.bus.local.LocalMessageBus">
 		<property name="queueSize"
 			value="${xd.local.transport.named.queueSize: #{T(Integer).MAX_VALUE}}" />
 		<property name="poller">
